Moved ICommand to base Command package
