+++ app/src/main/java/com/example/ryanblaser/tickettoride/Client/ClientCommunicator.java
@@ -3,6 +3,7 @@ package com.example.ryanblaser.tickettoride.Client;
 import android.os.AsyncTask;
 
 
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.Command.Phase1.*;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
