+++ app/src/main/java/com/example/ryanblaser/tickettoride/Client/ServerProxy.java
@@ -1,6 +1,7 @@
 package com.example.ryanblaser.tickettoride.Client;
 
 
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.Command.Phase1.AddGameToServerCommand;
 import com.example.ryanblaser.tickettoride.Command.Phase1.AddPlayerToClientCommand;
