+++ app/src/main/java/com/example/ryanblaser/tickettoride/Command/Phase1/AddGameToServerCommand.java
@@ -1,6 +1,7 @@
 package com.example.ryanblaser.tickettoride.Command.Phase1;
 
 import com.example.ryanblaser.tickettoride.Client.User;
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.Server.IServer;
 import com.example.ryanblaser.tickettoride.ServerModel.GameModels.Game;
