+++ app/src/main/java/com/example/ryanblaser/tickettoride/Command/Phase1/AddPlayerToClientCommand.java
@@ -1,5 +1,6 @@
 package com.example.ryanblaser.tickettoride.Command.Phase1;
 import com.example.ryanblaser.tickettoride.Client.ClientFacade;
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.Server.IServer.GameIsFullException;
 import com.example.ryanblaser.tickettoride.Client.User;
 import com.example.ryanblaser.tickettoride.ServerModel.GameModels.Game;
