+++ app/src/main/java/com/example/ryanblaser/tickettoride/Command/Phase1/AddResumableToClientCommand.java
@@ -2,6 +2,7 @@ package com.example.ryanblaser.tickettoride.Command.Phase1;
 
 import com.example.ryanblaser.tickettoride.Client.ClientFacade;
 import com.example.ryanblaser.tickettoride.Client.User;
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.ServerModel.GameModels.Game;
 import com.fasterxml.jackson.annotation.JsonIgnore;
