+++ app/src/main/java/com/example/ryanblaser/tickettoride/Command/Phase1/GetCommandsCommand.java
@@ -1,4 +1,5 @@
 package com.example.ryanblaser.tickettoride.Command.Phase1;
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.Server.ClientProxy;
 import com.example.ryanblaser.tickettoride.Client.User;
 import com.example.ryanblaser.tickettoride.ServerModel.GameModels.Game;
