+++ app/src/main/java/com/example/ryanblaser/tickettoride/Command/Phase2/SelectRequestedDestinationCardCommand.java
@@ -1,7 +1,7 @@
 package com.example.ryanblaser.tickettoride.Command.Phase2;
 
 import com.example.ryanblaser.tickettoride.Client.User;
-import com.example.ryanblaser.tickettoride.Command.Phase1.ICommand;
+import com.example.ryanblaser.tickettoride.Command.ICommand;
 import com.example.ryanblaser.tickettoride.Command.Phase1.CommandContainer;
 import com.example.ryanblaser.tickettoride.Server.IServer;
 import com.example.ryanblaser.tickettoride.ServerModel.GameModels.Game;
